requirements.txt
fastapi 
redis 
pyjwt 
uvicorn 
python-dotenv
pytest
prometheus-client